% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-dosage.R
\name{convert.allele.frequency.to.dosage}
\alias{convert.allele.frequency.to.dosage}
\title{Convert allele frequency to mean dosage}
\usage{
convert.allele.frequency.to.dosage(allele.frequency)
}
\arguments{
\item{allele.frequency}{A numeric vector of allele frequencies.}
}
\value{
A numeric vector of mean dosages for the allele frequencies.
}
\description{
Convert a population allele frequency to a mean dosage for that allele.
}
\examples{
allele.frequency <- seq(0.1, 0.9, 0.1);
convert.allele.frequency.to.dosage(allele.frequency);
}
