% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pgs.R
\name{create.pgs.density.plot}
\alias{create.pgs.density.plot}
\title{Plot PGS Density}
\usage{
create.pgs.density.plot(
  pgs.data,
  phenotype.columns = NULL,
  output.dir = NULL,
  filename.prefix = NULL,
  file.extension = "png",
  tidy.titles = FALSE,
  width = 10,
  height = 10,
  xaxes.cex = 1.5,
  yaxes.cex = 1.5,
  titles.cex = 1.5,
  key.cex = 1,
  border.padding = 1
)
}
\arguments{
\item{pgs.data}{data.frame PGS data with columns outputted by apply.polygenic.score()}

\item{phenotype.columns}{character vector of phenotype columns in pgs.data to plot (optional)}

\item{output.dir}{character directory to save output plots}

\item{filename.prefix}{character prefix for output filenames}

\item{file.extension}{character file extension for output plots}

\item{tidy.titles}{logical whether to reformat PGS plot titles to remove periods}

\item{width}{numeric width of output plot in inches}

\item{height}{numeric height of output plot in inches}

\item{xaxes.cex}{numeric cex for all x-axis labels in inches}

\item{yaxes.cex}{numeric cex for all y-axis labels in inches}

\item{titles.cex}{numeric cex for all plot titles in inches}

\item{key.cex}{numeric size of color key legend}

\item{border.padding}{numeric padding for plot borders}
}
\value{
if no output.dir is provided, a multipanel lattice plot object is returned, otherwise a plot is written to the indicated path and NULL is returned.
}
\description{
Plot density curves of PGS data from automatically detected PGS columns outputted by apply.polygenic.score (PGS, PGS.with.replaced.missing, PGS.with.normalized.missing).
If phenotype columns are provided, multiple density curves are plotted for automatically detected categories for each categorical variable.
}
\examples{
set.seed(200);
pgs.data <- data.frame(
    PGS = rnorm(200, 0, 1)
    );
 temp.dir <- tempdir();

# Basic Plot
create.pgs.density.plot(pgs.data, output.dir = temp.dir, filename.prefix = 'basic-plot');

# Plot multiple PGS outputs
pgs.data$PGS.with.normalized.missing <- rnorm(200, 1, 1);
create.pgs.density.plot(pgs.data, output.dir = temp.dir);

# Plot phenotype categories
pgs.data$sex <- sample(c('male', 'female', 200, replace = TRUE));
create.pgs.density.plot(
    pgs.data, output.dir = temp.dir,
    filename.prefix = 'multiple-pgs',
    phenotype.columns = 'sex'
    );

# Plot multiple phenotypes
pgs.data$letters <- sample(letters[1:5], 200, replace = TRUE);
create.pgs.density.plot(
 pgs.data, output.dir = temp.dir,
 filename.prefix = 'multiple-phenotypes',
 phenotype.columns = c('sex', 'letters')
 );
}
