% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pgs.R
\name{create.pgs.with.continuous.phenotype.plot}
\alias{create.pgs.with.continuous.phenotype.plot}
\title{Plot PGS Scatterplots}
\usage{
create.pgs.with.continuous.phenotype.plot(
  pgs.data,
  phenotype.columns,
  output.dir = NULL,
  filename.prefix = NULL,
  file.extension = "png",
  tidy.titles = FALSE,
  compute.correlation = TRUE,
  corr.legend.corner = c(0, 1),
  corr.legend.cex = 1.5,
  width = 10,
  height = 10,
  xaxes.cex = 1.5,
  yaxes.cex = 1.5,
  titles.cex = 1.5,
  point.cex = 0.75,
  border.padding = 1
)
}
\arguments{
\item{pgs.data}{data.frame PGS data with columns of continuous phenotype variables outputted by apply.polygenic.score()}

\item{phenotype.columns}{character vector of continuous phenotype columns in pgs.data to plot}

\item{output.dir}{character directory to save output plots}

\item{filename.prefix}{character prefix for output filenames}

\item{file.extension}{character file extension for output plots}

\item{tidy.titles}{logical whether to reformat PGS plot titles to remove periods}

\item{compute.correlation}{logical whether to compute correlation between PGS and phenotype and display in plot}

\item{corr.legend.corner}{numeric vector indicating the corner of the correlation legend e.g. c(0,1) for top left}

\item{corr.legend.cex}{numeric cex for correlation legend}

\item{width}{numeric width of output plot in inches}

\item{height}{numeric height of output plot in inches}

\item{xaxes.cex}{numeric cex for x-axis labels}

\item{yaxes.cex}{numeric cex for y-axis labels}

\item{titles.cex}{numeric cex for plot titles}

\item{point.cex}{numeric cex for plot points}

\item{border.padding}{numeric padding for plot borders}
}
\value{
multipanel plot object
}
\description{
Create scatterplots for PGS data with continuous phenotype variables
}
\examples{
set.seed(200);

pgs.data <- data.frame(
    PGS = rnorm(200, 0, 1),
    continuous.phenotype = rnorm(200, 2, 1)
    );
 temp.dir <- tempdir();

# Basic Plot
create.pgs.with.continuous.phenotype.plot(
    pgs.data,
    output.dir = temp.dir,
    filename.prefix = 'basic-plot',
    phenotype.columns = 'continuous.phenotype'
    );

# Plot multiple PGS outputs
pgs.data$PGS.with.normalized.missing <- rnorm(200, 1, 1);
create.pgs.with.continuous.phenotype.plot(
    pgs.data,
    output.dir = temp.dir,
    filename.prefix = 'multiple-pgs',
    phenotype.columns = 'continuous.phenotype'
    );

# Plot multiple phenotypes
pgs.data$continuous.phenotype2 <- rnorm(200, 10, 1);
create.pgs.with.continuous.phenotype.plot(
    pgs.data,
    output.dir = temp.dir,
    filename.prefix = 'multiple-phenotypes',
    phenotype.columns = c('continuous.phenotype', 'continuous.phenotype2')
    );
}
