% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-pgs-statistics.R
\name{get.pgs.percentiles}
\alias{get.pgs.percentiles}
\title{get.pgs.percentiles}
\usage{
get.pgs.percentiles(pgs, n.percentiles = NULL)
}
\arguments{
\item{pgs}{numeric vector of polygenic scores}

\item{n.percentiles}{integer number of percentiles to calculate (optional)}
}
\value{
data frame with columns for percentile, decile, quartile, and optional n.percentiles
}
\description{
Calculate percentiles and report decile and quartile ranks for a vector of polygenic scores
}
