% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-weight-files.R
\name{import.pgs.weight.file}
\alias{import.pgs.weight.file}
\title{Import PGS weight file}
\usage{
import.pgs.weight.file(input, use.harmonized.data = TRUE)
}
\arguments{
\item{input}{A character string indicating the path to the input file.}

\item{use.harmonized.data}{A logical indicating whether the file should be formatted to indicate harmonized data columns for use in future PGS application.}
}
\value{
A list containing the file metadata and the weight data.
}
\description{
Import a PGS weight file formatted according to PGS catalog guidelines, and prepare for PGS application.
}
