% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pgs.R
\name{plot.pgs.density}
\alias{plot.pgs.density}
\title{Plot PGS Density}
\usage{
plot.pgs.density(
  pgs.data,
  phenotype.columns = NULL,
  output.dir = NULL,
  filename.prefix = NULL,
  file.extension = "png",
  width = 10,
  height = 10,
  xaxes.cex = 1.5,
  yaxes.cex = 1.5,
  titles.cex = 1.5,
  border.padding = 1
)
}
\arguments{
\item{pgs.data}{data.frame with PGS data}

\item{phenotype.columns}{character vector of phenotype columns in pgs.data to plot (optional)}

\item{output.dir}{character directory to save output plots}

\item{filename.prefix}{character prefix for output filenames}

\item{file.extension}{character file extension for output plots}

\item{width}{numeric width of output plot in inches}

\item{height}{numeric height of output plot in inches}

\item{xaxes.cex}{numeric cex for x-axis labels in inches}

\item{yaxes.cex}{numeric cex for y-axis labels in inches}

\item{titles.cex}{numeric cex for plot titles in inches}

\item{border.padding}{numeric padding for plot borders}
}
\value{
multipanel plot object
}
\description{
Create density plots for PGS data
}
