% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pgs-rank.R
\name{plot.pgs.rank}
\alias{plot.pgs.rank}
\title{Plot PGS Rank}
\usage{
plot.pgs.rank(
  pgs.data,
  phenotype.columns = NULL,
  output.dir = NULL,
  filename.prefix = NULL,
  file.extension = "png",
  width = 8,
  height = 8,
  xaxis.cex = 1.2,
  yaxis.cex = 1,
  titles.cex = 1.2,
  border.padding = 1
)
}
\arguments{
\item{pgs.data}{data.frame PGS data containing the following columns: Indiv, percentile, decile, quartile, n.missing.genotypes, and optionally user-defined percentiles and phenotype covariates.
This function is designed to work with the output of the function apply.polygenic.score().}

\item{phenotype.columns}{character vector of column names in pgs.data containing phenotype covariates to plot as color bars. Default is NULL.}

\item{output.dir}{character directory path to write plot to file. Default is NULL in which case the plot is returned as lattice multipanel object.}

\item{filename.prefix}{character prefix for plot filename.}

\item{file.extension}{character file extension for plot file. Default is 'png'.}

\item{width}{numeric width of plot in inches.}

\item{height}{numeric height of plot in inches.}

\item{xaxis.cex}{numeric size of x-axis labels.}

\item{yaxis.cex}{numeric size of y-axis labels.}

\item{titles.cex}{numeric size of plot titles.}

\item{border.padding}{numeric padding around plot border.}
}
\value{
lattice multipanel object if output.dir is NULL, otherwise a file is written to output.dir.
}
\description{
Plot PGS percentile rank of each sample as a barplot, plot missing genotypes if any are present, plot corresponding decile and quartile markers as a heatmap, optionally plot phenotype covariates as color bars.
}
