% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-output-files.R
\name{write.apply.polygenic.score.output.to.file}
\alias{write.apply.polygenic.score.output.to.file}
\title{Write apply.polygenic.score output to file}
\usage{
write.apply.polygenic.score.output.to.file(
  apply.polygenic.score.output,
  output.dir,
  file.prefix = NULL
)
}
\arguments{
\item{apply.polygenic.score.output}{list of two data frames: pgs.output and regression.output}

\item{output.dir}{character string of the path to write both output files}

\item{file.prefix}{character string of the file prefix to use for both output files}
}
\description{
A utility function that writes the two data frames outputted by apply.polygenic.score to two tab-delimited text files.
}
