% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-output-files.R
\name{write.per.sample.pgs.table}
\alias{write.per.sample.pgs.table}
\title{Write PGS per sample table to file}
\usage{
write.per.sample.pgs.table(
  per.sample.pgs.summary.data,
  output.path,
  file.prefix = NULL
)
}
\arguments{
\item{per.sample.pgs.summary.data}{data.frame of PGS per sample data}

\item{output.path}{character string of the path to write the PGS per sample table}

\item{file.prefix}{character string of the file prefix to use for the PGS per sample table}
}
\description{
Write PGS per sample summary data table to tab separated text file.
}
