% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-pgs.R
\name{apply.polygenic.score}
\alias{apply.polygenic.score}
\title{Apply polygenic score to VCF data}
\usage{
apply.polygenic.score(
  vcf.data,
  pgs.weight.data,
  phenotype.data = NULL,
  phenotype.analysis.columns = NULL,
  output.dir = NULL,
  file.prefix = NULL,
  missing.genotype.method = "mean.dosage",
  use.external.effect.allele.frequency = FALSE,
  n.percentiles = NULL,
  analysis.source.pgs = NULL
)
}
\arguments{
\item{vcf.data}{A data.frame containing VCF genotype data as formatted by \code{import.vcf()}.}

\item{pgs.weight.data}{A data.frame containing PGS weight data as formatted by \code{import.pgs.weight.file()}.}

\item{phenotype.data}{A data.frame containing phenotype data. Must have an Indiv column matching vcf.data. Default is \code{NULL}.}

\item{phenotype.analysis.columns}{A character vector of phenotype columns from phenotype.data to analyze. Default is \code{NULL}.}

\item{output.dir}{A character string indicating the directory to write output files. Separate files are written for per-sample pgs results and optional regression results.
Files are tab-separate .txt files. Default is NULL in which case no files are written.}

\item{file.prefix}{A character string to prepend to the output file names. Default is \code{NULL}.}

\item{missing.genotype.method}{A character string indicating the method to handle missing genotypes. Options are "mean.dosage", "normalize", or "none". Default is "mean.dosage".}

\item{use.external.effect.allele.frequency}{A logical indicating whether to use an external effect allele frequency for calculating mean dosage when handling missing genotypes. Default is \code{FALSE}.
Provide allele frequency as a column is \code{pgs.weight.data} named \code{allelefrequency_effect}.}

\item{n.percentiles}{An integer indicating the number of percentiles to calculate for the PGS. Default is \code{NULL}.}

\item{analysis.source.pgs}{A character string indicating the source PGS for percentile calculation and regression analyses. Options are "mean.dosage", "normalize", or "none".
When not specified, defaults to missing.genotype.method choice and if more than one PGS missing genotype method is chosen, calculation defaults to the first selection.}
}
\value{
A list containing the PGS per sample and regression output if phenotype analysis columns are provided.

\strong{Output Structure}
The outputed list contains the following elements:
\itemize{
\item pgs.output: A data.frame containing the PGS per sample and optional phenotype data.
\item regression.output: A data.frame containing the results of the regression analysis if phenotype.analysis.columns are provided.
}

pgs.output columns:
\itemize{
\item Indiv: A character string indicating the sample ID.
\item PGS: A numeric vector indicating the PGS per sample. (only if missing.genotype.method includes "none")
\item PGS.with.normalized.missing: A numeric vector indicating the PGS per sample with missing genotypes normalized. (only if missing.genotype.method includes "normalize")
\item PGS.with.replaced.missing: A numeric vector indicating the PGS per sample with missing genotypes replaced by mean dosage. (only if missing.genotype.method includes "mean.dosage")
\item percentile: A numeric vector indicating the percentile rank of the PGS.
\item decile: A numeric vector indicating the decile rank of the PGS.
\item quartile: A numeric vector indicating the quartile rank of the PGS.
\item percentile.X: A numeric vector indicating the user-specified percentile rank of the PGS. (only if n.percentiles is specified)
\item n.missing.genotypes: A numeric vector indicating the number of missing genotypes per sample.
\item percent.missing.genotypes: A numeric vector indicating the percentage of missing genotypes per sample.
}

regression.output columns:
\itemize{
\item phenotype.analysis.columns: A character vector of phenotype columns analyzed.
\item beta: A numeric vector indicating the beta coefficient of the regression analysis.
\item se: A numeric vector indicating the standard error of the beta coefficient.
\item p.value: A numeric vector indicating the p-value of the beta coefficient.
\item r.squared: A numeric vector indicating the r-squared value of the regression analysis. NA for logistic regression.
\item AUC: A numeric vector indicating the area under the curve of the regression analysis. NA for linear regression.
}

\strong(PGS Calculation)
PGS for each individual is calculated as the sum of the product of the dosage and beta coefficient for each variant in the PGS:
}
\description{
Apply a polygenic score to VCF data.
}
\examples{
# Example VCF
vcf.path <- system.file(
    'extdata',
    'HG001_GIAB.vcf.gz',
    package = 'ApplyPolygenicScore',
    mustWork = TRUE
    );
vcf.import <- import.vcf(vcf.path);

# Example pgs weight file
pgs.weight.path <- system.file(
    'extdata',
    'PGS000662_hmPOS_GRCh38.txt.gz',
    package = 'ApplyPolygenicScore',
    mustWork = TRUE
    );
pgs.import <- import.pgs.weight.file(pgs.weight.path);

pgs.data <- apply.polygenic.score(
    vcf.data = vcf.import$dat,
    pgs.weight.data = pgs.import$pgs.weight.data,
    missing.genotype.method = 'none'
    );

# Specify different methods for handling missing genotypes
pgs.import$pgs.weight.data$allelefrequency_effect <- rep(0.5, nrow(pgs.import$pgs.weight.data));
pgs.data <- apply.polygenic.score(
    vcf.data = vcf.import$dat,
    pgs.weight.data = pgs.import$pgs.weight.data,
    missing.genotype.method = c('none', 'mean.dosage', 'normalize'),
    use.external.effect.allele.frequency = TRUE
    );

# Provide phenotype data for basic correlation analysis
phenotype.data <- data.frame(
    Indiv = unique(vcf.import$dat$Indiv),
    continuous.phenotype = rnorm(length(unique(vcf.import$dat$Indiv))),
    binary.phenotype = sample(
        c('a', 'b'),
        length(unique(vcf.import$dat$Indiv)),
        replace = TRUE
        )
    );

pgs.data <- apply.polygenic.score(
    vcf.data = vcf.import$dat,
    pgs.weight.data = pgs.import$pgs.weight.data,
    phenotype.data = phenotype.data
    );

}
