% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-pgs.R
\name{apply.polygenic.score}
\alias{apply.polygenic.score}
\title{Apply polygenic score to VCF data}
\usage{
apply.polygenic.score(vcf.data, pgs.weight.data)
}
\arguments{
\item{vcf.data}{A data.frame containing VCF genotype data.}

\item{pgs.weight.data}{A data.frame containing PGS weight data.}
}
\value{
A data.frame containing the PGS per sample.
}
\description{
Apply a polygenic score to VCF data.
}
