% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-pgs.R
\name{apply.polygenic.score}
\alias{apply.polygenic.score}
\title{Apply polygenic score to VCF data}
\usage{
apply.polygenic.score(
  vcf.data,
  pgs.weight.data,
  missing.genotype.method = "mean.dosage",
  use.external.effect.allele.frequency = FALSE,
  n.percentiles = NULL
)
}
\arguments{
\item{vcf.data}{A data.frame containing VCF genotype data.}

\item{pgs.weight.data}{A data.frame containing PGS weight data.}

\item{missing.genotype.method}{A character string indicating the method to handle missing genotypes. Options are "mean.dosage", "normalize", or "none". Default is "mean.dosage".}

\item{use.external.effect.allele.frequency}{A logical indicating whether to use an external effect allele frequency for calculating mean dosage when handling missing genotypes. Default is FALSE.}

\item{n.percentiles}{An integer indicating the number of percentiles to calculate for the PGS. Default is NULL.}
}
\value{
A data.frame containing the PGS per sample.
}
\description{
Apply a polygenic score to VCF data.
}
