% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-pgs.R
\name{apply.polygenic.score}
\alias{apply.polygenic.score}
\title{Apply polygenic score to VCF data}
\usage{
apply.polygenic.score(
  vcf.data,
  pgs.weight.data,
  phenotype.data = NULL,
  phenotype.analysis.columns = NULL,
  output.dir = NULL,
  file.prefix = NULL,
  missing.genotype.method = "mean.dosage",
  use.external.effect.allele.frequency = FALSE,
  n.percentiles = NULL,
  analysis.source.pgs = NULL
)
}
\arguments{
\item{vcf.data}{A data.frame containing VCF genotype data.}

\item{pgs.weight.data}{A data.frame containing PGS weight data.}

\item{phenotype.data}{A data.frame containing phenotype data. Must have an Indiv column matching vcf.data. Default is NULL.}

\item{phenotype.analysis.columns}{A character vector of phenotype columns to analyze. Default is NULL.}

\item{output.dir}{A character string indicating the directory to write output files. Separate files are written for per-sample pgs results and optional regression results.
Files are tab-separate .txt files. Default is NULL in which case no files are written.}

\item{file.prefix}{A character string to prepend to the output file names. Default is NULL.}

\item{missing.genotype.method}{A character string indicating the method to handle missing genotypes. Options are "mean.dosage", "normalize", or "none". Default is "mean.dosage".}

\item{use.external.effect.allele.frequency}{A logical indicating whether to use an external effect allele frequency for calculating mean dosage when handling missing genotypes. Default is FALSE.}

\item{n.percentiles}{An integer indicating the number of percentiles to calculate for the PGS. Default is NULL.}

\item{analysis.source.pgs}{A character string indicating the source PGS for percentile calculation and regression analyses. Options are "mean.dosage", "normalize", or "none".
When not specified, defaults to missing.genotype.method choice and if more than one PGS missing genotype method is chosen, calculation defaults to the first selection.}
}
\value{
A list containing the PGS per sample and regression output if phenotype analysis columns are provided.
}
\description{
Apply a polygenic score to VCF data.
}
