% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-weight-files.R
\name{check.pgs.weight.columns}
\alias{check.pgs.weight.columns}
\title{Check PGS weight file columns}
\usage{
check.pgs.weight.columns(pgs.weight.colnames, harmonized = TRUE)
}
\arguments{
\item{pgs.weight.colnames}{A character vector of column names.}

\item{harmonized}{A logical indicating whether the presence of harmonized columns should be checked.}
}
\value{
A logical indicating whether the file contains the required columns.
}
\description{
Check that a PGS weight file contains the required columns.
}
\examples{
pgs.weight.colnames <- c(
    'chr_name',
    'chr_position',
    'effect_allele',
    'effect_weight'
    );
check.pgs.weight.columns(c(pgs.weight.colnames, 'hm_chr', 'hm_pos'));

check.pgs.weight.columns(
    pgs.weight.colnames,
    harmonized = FALSE
    );
}
