% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-vcf.R
\name{import.vcf}
\alias{import.vcf}
\title{Import VCF file}
\usage{
import.vcf(vcf.path, info.fields = NULL, format.fields = NULL, verbose = FALSE)
}
\arguments{
\item{vcf.path}{A character string indicating the path to the VCF file to be imported.}

\item{info.fields}{A character vector indicating the INFO fields to be imported.}

\item{format.fields}{A character vector indicating the FORMAT fields to be imported.}

\item{verbose}{A logical indicating whether verbose output should be printed by vcfR.}
}
\value{
A long form data.frame containing the VCF data.
}
\description{
Import a VCF file into a long form data.frame using vcfR package.
}
