% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pgs.R
\name{plot.pgs.with.continuous.phenotype}
\alias{plot.pgs.with.continuous.phenotype}
\title{Plot PGS Scatterplots}
\usage{
plot.pgs.with.continuous.phenotype(
  pgs.data,
  phenotype.columns,
  output.dir = getwd(),
  filename.prefix = NULL,
  file.extension = "png",
  width = 10,
  height = 10,
  xaxes.cex = 1.5,
  yaxes.cex = 1.5,
  titles.cex = 1.5
)
}
\arguments{
\item{pgs.data}{data.frame with PGS and phenotype data}

\item{phenotype.columns}{character vector of continuous phenotype columns in pgs.data to plot}

\item{output.dir}{character directory to save output plots}

\item{filename.prefix}{character prefix for output filenames}

\item{file.extension}{character file extension for output plots}

\item{width}{numeric width of output plot in inches}

\item{height}{numeric height of output plot in inches}

\item{xaxes.cex}{numeric cex for x-axis labels}

\item{yaxes.cex}{numeric cex for y-axis labels}

\item{titles.cex}{numeric cex for plot titles}
}
\value{
multipanel plot object
}
\description{
Create scatterplots for PGS data with continuous phenotype variables
}
